+++ tests/Fixtures/Author.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace JacobJoergensen\LaravelPaper\Tests\Fixtures;
 
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Collection;
 use JacobJoergensen\LaravelPaper\Attributes\ContentPath;
 use JacobJoergensen\LaravelPaper\Attributes\Driver;
 use JacobJoergensen\LaravelPaper\Paper;
+use JacobJoergensen\LaravelPaper\Relations\HasManyPaper;
 
 #[Driver('json')]
 #[ContentPath('tests/content/authors')]
@@ -16,7 +16,7 @@ final class Author extends Model
 {
     use Paper;
 
-    public function posts(): Collection
+    public function posts(): HasManyPaper
     {
         return $this->hasManyPaper(Post::class);
     }